Move the attrs handling code into it's own module.
